kernel: Boost whenever a zygote-forked process becomes a top app
Boost the CPU to the max for 1000 ms whenever the top app changes, which improves app launch speeds and addresses jitter when switching between apps. A check to make sure that the top-app's parent is zygote ensures that a user-facing app is indeed what's added to the top app task group, since app processes are forked from zygote.

TenSeventy7: In our case, we don't really need the CPU boost as the userspace already handles it for us, also because we don't have the CPU input boost driver. Just boost devfreq early so such CPU boost comes in time.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: John Vincent <git@tenseventyseven.cf>