arm64: defconfig: Set "cfq" as default I/O scheduler
So I experienced a spontaneous breakage of an SD card after trying to move an app from the internal storage to the SD card. Assuming it was a kernel issue, I tried reformatting the card on various devices. But they don't work as well.

It might be because that the SD card was already in the verge of breaking (please put S.M.A.R.T. support in these, god) but testing with 'cfq' on a different SD card on the same device works, and I'm thinking it might be because of 'maple' being set as I/O schedule, which I personally don't really think is the cause.

But to be safe, set CFQ back as default. We'll optimize this further in future commits.

Signed-off-by: John Vincent <git@tenseventyseven.cf>