ems: freqvar_tune: Block limit changes to `schedutil`
This pretty much makes configuring it from Kconfig, as well as on sysfs useless as DT will just override it. Since we want to make it possible without modifying DT, just block it from changing rate limits.

It shouldn't be much of an issue since it's pretty much '5000' on all frequencies from DT. Imagine adding this feature but not even using it properly lol

Signed-off-by: John Vincent <git@tenseventyseven.cf>