tools: package: Enable experimental support for native ZRAM writeback on AOSP ROMs.
This adds experimental ZRAM writeback support to non-One UI ROMs/AOSP ROMs and GSIs. Android has a native ZRAM writeback that, while not as aggressive as RAM Plus, should improve performance if configured properly.

We'll be following the Pixel's 1/24/24 ZRAM writeback rule for this, and make an overlay to be installed in system_ext, which should work on both AOSP and GSIs.

Signed-off-by: John Vincent <git@tenseventyseven.cf>