BACKPORT: melfas: melfas_mss100: Inline model-dependent functions directly
Overall UI/UX performance was noticed with this change. Adapt to the (older) drivers used on the Samsung A50 and A50S.

Instead of inlining mms_lowpower_mode as an 'extern' function, put them directly inside melfas_mss100.c instead. Clang errors out with 'undefined reference' errors when LTO is enabled.

Thanks to these guys for helping with this cleanup.

Co-authored-by: AnuragRai <kerneldeveloper7232@gmail.com>
Co-authored-by: Dakkshesh <dakkshesh5@gmail.com>
Co-authored-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: John Vincent <git@tenseventyseven.cf>