sched: Consider the original capacity of the CPU for margin calculation
The scheduler (including EMS) from Samsung's 4.19 kernels calculates a task's margins based on the CPU's original capacity, not just 'SCHED_CAPACITY_SCALE'. Test this change on ours if we get better results on scheduling.

Instead of using capacity_orig_of(), use EMS' get_cpu_max_capacity() to ensure that we get the actual max capacity of the CPU in question.

Signed-off-by: John Vincent <git@tenseventyseven.cf>