arm64: dts: exynos: Re-adjust variable frequency boosting (freqvar-boost)
Now that upstream patches and improvements to the scheduler have been applied, it's also important to address the variable frequency scaling the kernel does. Also known as 'freqvar-tune', it adjusts the CPU frequency by inflating the reported utilization to both schedutil (upscale_ratio_table) and schedtune (boost_table), making the CPU boost to higher frequencies. This can be treated as some sort of "wiggle room" just in case a task suddenly adds a significant amount of load.

Based on device testing with the current boosting values, there is aggressive high-frequency scaling on the little cores as well as measly frequency scaling on the big cores. Not only does this cause performance issues, it's also troublesome for battery life.

Re-adjust these values again to adapt to the new scheduler changes, based on additional references, and to improve overall device performance.

References:
https://github.com/CruelKernel/samsung-exynos9820/blob/cruel-HVE5-v5.2/arch/arm64/boot/dts/exynos/exynos9610.dtsi
https://github.com/TheGalaxyProject/tgpkernel-n9-p/blob/master/arch/arm64/boot/dts/exynos/exynos9810.dtsi

Signed-off-by: John Vincent <git@tenseventyseven.cf>