arm64: dts: exynos: Re-adjust Fluid RT (FRT) ratios
Samsung's Fluid RT (FRT) FRT aims to place real-time (RT) tasks to CPUs that meet the 'coverage' and 'active' ratio requirements during placement. Coverage ratio is used to find the first CPU with RT usage (other tasks not accounted for) below 'coverage-ratio', and active ratio is used to find candidate coregroups (clusters) based on the cluster's total usage (including non-RT tasks).

It makes RT task boosting much more complicated than what it's supposed to be, but it is what it is. It looks good in hindsight, but doesn't really work well in practice. It would have been nice as this makes the RT scheduler capacity-aware, but the way Samsung implemented it is just not great.

Increase the coverage to allow CPUs with 30% RT usage on little and half of it on the big cluster to get placed RT tasks per-core, and 50% and 25% total usage on the little and big clusters respectively for placing RT tasks on a coregroup basis. The big cluster has larger capacity but can be battery-intensive so do it this way.

Signed-off-by: John Vincent <git@tenseventyseven.cf>