sched: core: Only apply schedtune boost to RT tasks
While it's nice to boost 50ms on all task classes, this can hurt a bit on battery life. We already enabled iowait boosting on all performance presets on FreshROMs, as well as placing the top-app's RenderThread and UI threads as RT tasks. It makes much more sense to just have this disabled, as-is the default. We still get boosting on RT tasks but not on less important ones.

This partially reverts commit 927c3540a4a79925af47687871523b8ecff374c2.

Signed-off-by: John Vincent <git@tenseventyseven.cf>