arm64: dts: exynos: Experiment with lowering the little cluster's over-utilization ratio
Samsung has recently changed the scheduler's "overutilized" ratio from 30% on the little cluster and 40% on the big cluster to 80% on both clusters. With the changes to the Exynos Mobile scheduler, we might actually benefit from slightly-stricter load balancing on coregroup0 (little cores).

Follow the "ratio-10/ratio" rule Samsung used in the past by decreasing the little cluster's overutil ratio to 70%. This shouldn't really change much considering the little cores have not-so-huge capacities, but it's nice to have some wiggle room for load-balancing on the core that handles too many tasks.

Signed-off-by: John Vincent <git@tenseventyseven.cf>