sched: ems: Only mitigate preemption if CPU has a boosted task
The commit that inspired changes to 'proper CPU' selection[1], af8c499 ("sched: ems: Avoid non-latency-sensitive task preempting latency-sensitive tasks") was based on a commit that initially aimed to mitigate tasks that aren't 'boosted' to preempt CPUs with those who do[2].

The problem with "schedtune.prefer_perf", at least in FreshROMs' case, is that we use it as a 'placement hint' to place tasks (notably "top-app" tasks) in performant cores. A number of tasks in 'top-app' aren't necessarily latency-sensitive; and while it prefers the big cluster, tasks don't necessarily get placed in there. Like this example of a device in the home screen:

PSR   PID NAME
  5  4507 zygote64
  5  4508 zygote
  6  4661 cameraserver
  0  4966 system_server
  1  5637 com.android.systemui
  6  8675 com.google.android.googlequicksearchbox:search
  6 11054 com.sec.android.app.launcher

This can pose a problem as all of these processes are marked "prefer-perf". And based on observation, these get migrated and scheduled on different cores from time-to-time. This is a much bigger issue on standby, where:

PSR   PID NAME
  4  4507 zygote64
  5  4508 zygote
  6  4661 cameraserver
  3  4966 system_server
  2  5637 com.android.systemui
  1 11054 com.sec.android.app.launcher

3/4 of our big cluster cores are placed with prefer-perf tasks. While not a problem on the little cluster (because of 605006c), it can cause excessive task packing on the big cluster as the majority of the big cluster have "prefer-perf" tasks. This causes performance regressions from time-to-time, especially when the device is experiencing high load (i.e. post-standby because of ea9943c).

[2] initially used the 'schedtune.boost' hint for this, before being replaced by a newer 'schedtune.prefer_high_cap' hint[1]. Using 'schedtune.boost' in this case might actually be good for us, since the placement hint used by 'boost' (to place tasks on the big cluster) is already handled by 'schedtune.prefer_perf' in Exynos' scheduler. We only use the 'boost' hint on app launch as well as on the task currently in UX (see e538203), cases where we actually want preemption to be avoided as it causes UI jank.

Since 'schedtune.boost' pretty much means very little at this point, we can take advantage of this to completely separate these placement hints, as is the case with [1]. Instead of checking for prefer_perf, check for boost instead when checking for preemption.

With this patch merged, there is better scheduling for top-app tasks with 'prefer-perf':

PSR   PID ARGS
  5  4506 zygote64
  5  4507 zygote
  5  4674 cameraserver
  2  4975 system_server
  0  5622 com.android.systemui
  7  5927 com.sec.android.app.launcher
  6  8491 com.google.android.googlequicksearchbox:search
  6 12622 su

[1]: https://android.googlesource.com/kernel/msm/+/cfee90b
[2]: https://android.googlesource.com/kernel/msm/+/da7768c

Signed-off-by: John Vincent <git@tenseventyseven.cf>