sched: ems: Only mitigate preemption if the CPU has a 'kpp' task
The problem with "schedtune.prefer_perf", at least in our case, is that we use it as a 'placement hint' to place tasks (notably the UX task) on performant cores. But they aren't necessarily something we want to avoid preemption for every time as this can cause excessive task packing on the big cluster as there are at most two "prefer-perf" tasks, namely the UX task and the System UI. This causes performance regressions from time-to-time, especially when the device is experiencing high load (i.e. post-standby because of ea9943c).

Google uses MSM's 'schedtune.prefer_high_cap' only during app launch as well as interaction, something that triggers the kernel's own "prefer_perf" hint. We can take advantage of this (and bring the exact same behavior as Google's) by only considering the kernel's prefer-perf (kpp_status) indicator when checking for preemption.

Signed-off-by: John Vincent <git@tenseventyseven.cf>