sched: ems: Change 'slowest' indicator to 'minimum capacity'
'is_slowest_cpu()' isn't intended to look for the "slowest" CPU per se, as the entire little cluster has exactly the same capacities on all of its cores. Instead, this is intended to be used to determine if the given CPU is in a cluster with the minimum capacity, i.e. the little cluster. It doesn't mean much for our platform (which is just bi-cluster), but looking at EMS it seems to support tri-cluster architectures as well.

Make the method clearer and change the method name to reflect its real purpose. There is no change in functionality. Naming is inspired by how MSM does theirs.

Signed-off-by: John Vincent <git@tenseventyseven.cf>