sched: ems: Optimize idle task placement and CPU selection
Exynos Mobile Scheduler (EMS) has its own placement logic for tasks that are 'prefer-idle', which is somehow very rudimentary in contrast to CFS. We cannot really use CFS' as exynos_wakeup_balance() almost always returns a valid CPU mask, so we are stuck with this. As with the newest improvements to EMS, improve selection on this area as well.

Similar to CFS, consider the CPU with the highest spare capacity over the one with lower utilization. Also make sure that the selected idle CPU is the shallowest, in order to reduce power consumption. For CPUs with the same cstate, prefer the one with lower utilization; based on work by Diep Quynh (<remilia.1505@gmail.com>).

Signed-off-by: John Vincent <git@tenseventyseven.cf>