sched: ems: Always favor CPUs with smaller capacity for proper CPU selection
At this point of the placement process, we have already done EAS scheduling as well as other placement hints. We just need to place this task in a CPU with a smaller capacity as this is not latency-sensitive.

Just like CFS, favor CPUs with smaller capacity and the highest spare utilization.

Signed-off-by: John Vincent <git@tenseventyseven.cf>