sched: ems: service: Use the estimated utilization of the task on placement
Follow what EMS uses for calculating a task's utilization on other placement paths such as 'ontime' and 'prefer-idle' as well as load-balancing (proper) CPU selection, and use the estimated utilization usage of a task before placing it.

Tasks that are 'prefer-perf' typically use a significant amount of CPU in a short period of time (i.e. during app launch or game load) so it's good to have some wiggle room for task placement here.

Signed-off-by: John Vincent <git@tenseventyseven.cf>