sched: tune: Add an API to check solely for the kernel's 'prefer-perf' hint
We use 'schedtune.prefer_perf' as a placement hint to place tasks (such as the task currently in the UX or the System UI if it's currently in view) to a performant cluster, but they are not necessarily something we want to avoid preemption every time as this can cause slight performance issues in some cases, especially on heavy load.

There are instances however where the kernel's own 'prefer-perf' indicator can be useful, as this is set during app launch, device wake, and interaction; similar to how Google uses MSM's 'prefer-high-cap' hint[1]. Add an API to check solely for the kernel's own 'prefer-perf' hint. We'll use this for better preemption decisions.

[1]: https://android.googlesource.com/device/google/redbull/+/refs/heads/android12L-s1-release/powerhint.json

Signed-off-by: John Vincent <git@tenseventyseven.cf>