sched: tune: Add an API to check solely for the kernel's 'prefer-perf' hint
We use the 'prefer-perf' hint on stune as a placement hint to place tasks (such as top-app tasks) to a performant cluster, but they are not necessarily latency-sensitive (see 0be4fa0). There are instances however where the kernel's own 'prefer-perf' indicator can be useful, as this is set during app launch and device wake.

Add an API to check solely for the kernel's own 'prefer-perf' hint. We'll use this for better preemption decisions.

Signed-off-by: John Vincent <git@tenseventyseven.cf>