sched: ems: ontime: Improve 'ontime' CPU selection
Ontime tasks are both performance and latency-sensitive. Ensure that we place these tasks on CPUs with minimum utilization as much as possible.

For idle CPUs, ensure that we both get the CPU with the shallowest idle state (for fast wake and less power consumption) as well as the lowest utilization. Otherwise, use the estimated utilization of the task instead of its exact utilization so we have allowance for migration.

Signed-off-by: John Vincent <git@tenseventyseven.cf>