sched: tune: Extend schedtune_ux_task() to cover the 'prefer-perf' hint
Exynos Mobile Scheduler (EMS) uses the 'prefer-perf' hint to place tasks onto high-performance CPUs, as this typically means that the task is performance and/or latency-sensitive. Typically on CFS, only the 'schedtune.boost' hint is used for this.

We have changed EMS' 'prefer-idle' behavior to prefer lower capacity CPUs as much as possible to maintain good battery life and the UX task seems to get placed in this logic from time-to-time, making its performance not so great.

Let's extend schedtune_ux_task() for Exynos' 'prefer-perf' hint as well so we ensure that the task currently in the UX are on the big cluster cores as much as possible to maintain optimal performance without bringing miscellaneous tasks with it.

Signed-off-by: John Vincent <git@tenseventyseven.cf>