HACK: sched: tune: Allow the System UI to be a "UX task"
I'm just sick of the QS panel just jittering on swipe down because the (ugly) scheduler decides to migrate it to the little cluster despite it being already in the UX. Just put the entire thing on the big cluster and avoid preemption by making it a "UX task", even if its 'oom_score_adj' is below 0 (typically '-13').

This shouldn't be an issue for the most part as this doesn't skip the 'top-app' check, and will only do this if the SystemUI is a 'top app'. On One UI, this is only the case while the quick settings panel or the lock screen is shown. This is also the case when the device is on standby (presumably because of the lock screen), but it's set to have the "restricted" cpuset in that case, which locks the System UI on the little cluster.

This patch is very similar to Oplus' own UIFirst hacks, but very rudimentary in comparison.

Signed-off-by: John Vincent <git@tenseventyseven.cf>