sched: tune: Extend the 'UX task' metric to cover important, "UI-backer" threads as well
After observing the scheduler's behavior through 'ftrace' (tysm Samsung for adding tracing paths for EMS), it was found out that the cause of the UI jitter in userspace is caused by multiple things:

1. The UX task's 'RenderThread', which is mostly scheduled in the real-time (RT) path (on FreshROMs) is almost always terribly placed on the little cluster because it usually has the idlest cores. This is problematic as its very latency-sensitive, especially as an RT task. UI jitter is 'very' noticeable if it stalls, even for just a little bit.
2. Other threads used in UI processing, such as Mali's render backend (mali-cmar-backe'nd') are also placed the same way as RenderThread, making latency much worse.
3. These threads, if not considered RT, aren't given priority by EMS and are placed using the 'prefer-idle' or 'proper CPU' paths, which prefers CPUs of lower capacity.
4. The SystemUI (com.android.systemui) are placed the same way as 1-3, despite the QS panel or the lock screen being in the UX already, causing jitter when pulling the QS panel or swiping from the lock screen.

While pretty much a hack, let's extend our existing schedtune_adj_ta() to cover these kernel threads, as well as the System UI if it's currently shown to the user. Add an 'affine' parameter, which will be used by the scheduler to selectively apply priorities on the UX task alone, or including the backend threads running the UI. We'll use this for differentiating between task scheduling and preemption checking.

To ensure that we are not wasting CPU cycles or battery life on unimportant UX threads, ensure that we are only affining threads whose group leaders are at oom_score_adj of zero; typically the UX task's threads.

This is a strategy known to be used by Oppo and/or Oplus for their "UIFirst" feature. There's much more going on for them, but we only need the scheduling priorities for now.

Signed-off-by: John Vincent <git@tenseventyseven.cf>