mm: oom_kill: Only reap memory when screen is off
92d86a2 ("mm: oom_kill: reap memory of a task that receives SIGKILL") enabled the kernel to immediately reap the memory of a task that receives a SIGKILL in order to reclaim memory much quicker and on time. This however causes a bit of a performance hit[1], especially when it reaps a number of processes, which happens pretty often in Android.

However, we still do want this functionality as several builds of One UI (mostly ROMs based on firmware of flagship Samsung devices, such as FreshROMs) are lousy in reclaiming and even memory compaction, making the ROM consume a lot more memory than necessary.

Hook this functionality to the state notifier and only do memory reaping when the display is off or if the device is on standby. This ensures that we are only doing this kind of housekeeping when there is no activity currently going on in the foreground. We get optimal performance on activity, and good memory usage every time the device wakes up from standby.

[1]: https://github.com/cyberknight777/dragonheart_kernel_oneplus_sm8150/commit/3080795355dd93ded81f44fa6fc035dbe01ba284

Signed-off-by: John Vincent <git@tenseventyseven.cf>