sched: ems: Separate the 'high capacity' hint from 'performance-preferred'
We're encountering a lot of heavy UI stuttering if we place the UX task with the 'prefer-perf' hint as preemption is always avoided on the CPU where the UX task currently is. The problem is, we don't want 'top-app' to be only using the 'prefer-idle' hint as we want to put these tasks onto performant cores as much as possible.

Introduce a new 'prefer-high-cap' hint, inspired from Google's changes on the Pixel, instead for checking when the task either wants a high-capacity or a performant CPU. This way, we can still take advantage of the PCF for actual tasks then use this new hint for 'service tasks' which are largely latency-sensitive.

Also move the kernel's own 'prefer-perf' hint to this, as this is triggered on device wake, interaction, and app launch. This is also what Google uses the 'prefer-high-cap' hint for[1].

[1]: https://android.googlesource.com/device/google/redbull/+/refs/heads/android12L-s1-release/powerhint.json

Signed-off-by: John Vincent <git@tenseventyseven.cf>