sched: ems: prefer-idle: Ensure 'boosted' tasks are on higher-capacity CPUs
We want the UX task (app in the front of the user) to be placed on a higher-capacity CPU as much as possible as this is performance-sensitive and UI stutters are noticeable if it gets placed onto lower-capacity cores. While the UX task, being a 'top-app', is primarily placed using the 'prefer-perf' hint, there are instances where the PCF is unable to find a suitable CPU and placement is now done using the 'prefer-idle' hint.

For these tasks, ensure that we are placing them onto higher-capacity CPUs to ensure optimal performance.

Signed-off-by: John Vincent <git@tenseventyseven.cf>