sched: ems: Remove 'foreground' kernel-prefer-perf boosting on interaction and app launch
There's just so many background and miscellaneous tasks running on the 'foreground' schedtune group that we couldn't afford to avoid preemption on cores with these tasks every time the device wakes up, opens an app, or even on interaction. This causes excessive task packing as there's just at most 50 of them at a time, and avoiding preemption for all of them on the big cluster is just too much for the scheduler (and the little cluster) to handle.

Not only that, we're placing all of these on performant cores at almost every time through the 'prefer_perf_service' placement path on EMS, which is not great for battery life. Tasks in this schedtune group are not latency nor performance-sensitive, and at most just prefers to be on idle CPUs.

Remove 'kernel-prefer-perf' requests for the 'foreground' schedtune group on interaction (through Samsung's "input booster") as well as on CPU boosting (through 'exynos-ufc') so that they won't get unnecessary special treatment to be placed on performant cores as well as avoiding preemption on cores with these tasks in.

Signed-off-by: John Vincent <git@tenseventyseven.cf>