sched: ems: service/pcf: Avoid placing tasks on CPUs under 'ontime' migration
Tasks placed on the 'service' and 'prefer-perf' paths are performance-sensitive and we want them to be placed on CPUs that, while not necessarily higher-capacity (as is PCF), are not doing a bunch of migration to in order to reduce load balancing especially immediately after placing tasks on target CPUs.

Inspired by a tweak from MSM kernels[1], avoid placing tasks on CPUs where migration from 'ontime', typically heavy task migrations, are currently happening. This allows these tasks to be placed on CPUs that, while has more utilization, are at least not doing something as heavy as task migration.

[1]: https://github.com/cyberknight777/dragonheart_kernel_oneplus_sm8150/blob/d93d031a87c3401b5c5dfad104c98f12c1d6aca4/kernel/sched/fair.c#L7536

Signed-off-by: John Vincent <git@tenseventyseven.cf>