sched: ems: ontime: Directly get CPU cstate for comparison
The idle CPU selection logic for 'ontime' tasks fall apart when it couldn't get the CPU idle state struct as it immediately just gives the current CPU as the 'best' if it couldn't get its state. Use idle_get_state_idx(), something we've been using for other placement paths, to get the exit latency of the CPU so we still get the lowest cstate and utilization.

Signed-off-by: John Vincent <git@tenseventyseven.cf>