sched: pcf: Immediately put best idle CPU if a higher-capacity CPU is found
Re-initializing the best idle state and utilization metrics then going ahead with comparison (that won't happen anyway) can get expensive for a placement path that is quickly needed by performance and latency-sensitive tasks.

Instead of resetting the 'best' metrics, just go ahead and immediately put the current CPU as well as its cstate and utilization to be the 'best' then proceed with the next one.

Signed-off-by: John Vincent <git@tenseventyseven.cf>