sched: ems: service: Place important UX threads using the 'service' path
UX threads such as the 'RenderThread', even if are not RT, are latency-sensitive and causes UI jitter if not placed on a performant core. In addition, Android's display threads such as 'android.ui' and 'android.display' are biased from the userspace to be placed onto performant cores; let's clamp it further here.

Ensure the UX task, as well as its accompanying backends, are placed using the 'service' path to ensure that these threads are of optimal performance.

Signed-off-by: John Vincent <git@tenseventyseven.cf>