sched: frt: Always prefer higher-capacity CPUs for UX threads
Extend a74abc3 ("sched: frt: Initial prefer_perf adaption for Fluid RT") to also require all UX tasks, including the UX task's backing threads, to be placed on higher-capacity CPUs as much as possible. This ensures that we are getting the lowest latency for Android's UI and display threads, as well as the task's RenderThread and its Mali backend.

Signed-off-by: John Vincent <git@tenseventyseven.cf>