sched: ems: Remove 'task banding' scheduler path
Ultimately went unused on the Exynos 9610 and 9820, and is severely incomplete (which is probably why this went unused). Also tried to make this work but I just can't. Not to mention Exynos also gave up on this on 4.19.

Remove this unnecessary overhead and just remove it entirely from the tree.

This reverts commit 1d2290605f4b665d5826f3e660c0a4e9e44866d1.

Signed-off-by: John Vincent <git@tenseventyseven.cf>