kernel: configs: Make MGLRU optional for One UI builds
I'm getting anxious of over 0.5GB (500MB) less RAM on One UI because it likes to 'cache' apps and consume a lot of RAM. The problem is that One UI ROMs (even stock) have arbitrarily-added memory management parameters based on free RAM and thrashing is just too much if we go beyond 1800MB free RAM on a 6GB RAM variant Galaxy A50.

AOSP ROMs seem to play well with MGLRU so we're keeping it, and making it only enabled by default for AOSP builds. MGLRU is still present on One UI builds but it *must* be enabled by the user, or by the ROM developer if he/she/they are confident enough on their own memory management set in their ROMs.

Also enable Oplus' MM hacks on One UI ROMs since we're missing MGLRU and the ~200MB free RAM we net with this (when it doesn't conflict with MGLRU) is neat for our use case.

Signed-off-by: John Vincent <git@tenseventyseven.cf>